Add script to generate the registry cache tarball
To use the npm cache entirely, both the packages (tarballs) and registry
API responses must be cached. Running this script for a given package
will generate a tarball that can be unpacked in a spec file to the npm
cache dir and should allow it to work entirely offline. It should be run
on the oldest version of npm needed.